requirements.txt
django-environ==0.11.2
asgiref==3.7.2 
django==5.0.3 
sqlparse==0.4.4 
tzdata==2024.1
mysqlclient==2.2.4
amqp==5.2.0 
billiard==4.2.0 
celery==5.3.6 
click==8.1.7 
click-didyoumean==0.3.0 
click-plugins==1.1.1 
click-repl==0.3.0 
colorama==0.4.6 
kombu==5.3.5 
prompt-toolkit==3.0.43 
python-dateutil==2.9.0.post0 
redis==5.0.3 
six==1.16.0 
vine==5.1.0 
wcwidth==0.2.13
certifi==2024.2.2 
charset-normalizer==3.3.2 
idna==3.6 
requests==2.31.0 
urllib3==2.2.1
djangorestframework==3.15.0
django-cors-headers==4.3.1
drf-yasg==1.21.7 
inflection==0.5.1 
packaging==24.0 
pytz==2024.1 
pyyaml==6.0.1 
uritemplate==4.1.1
djangorestframework_simplejwt==5.3.1 
pyjwt==2.8.0
cffi==1.16.0 
cryptography==42.0.5
msal==1.28.0
pycparser==2.21
oauthlib==3.2.2 
requests_oauthlib==2.0.0